# docs/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Casa da Paz API
  version: 0.1.0
servers:
  - url: /api
paths:
...
  /auth/me:
    get:
      security: [{ bearerAuth: [] }]
      summary: Usuário atual
  /auth/usuarios:
    get:
      security: [{ bearerAuth: [] }]
      summary: Listar usuários (DIRETORIA)
    post:
      security: [{ bearerAuth: [] }]
      summary: Criar usuário
...
  /cobrancas:
    get:
      security: [{ bearerAuth: [] }]
      summary: Listar cobranças Asaas
    post:
      security: [{ bearerAuth: [] }]
      summary: Criar cobrança avulsa
  /cobrancas/assinaturas:
    get:
      security: [{ bearerAuth: [] }]
      summary: Assinaturas médiuns
    post:
      security: [{ bearerAuth: [] }]
      summary: Criar assinatura mensal
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT